requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
croniter==2.0.5
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
imageio-ffmpeg==0.4.9
multidict==6.0.5
numpy==1.26.4
py-cord==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
yarl==1.9.4
